+++ docs/conf.py
@@ -85,7 +85,7 @@ else:
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = ['_build', 'contracts', 'types', 'examples', 'grammar']
+exclude_patterns = ['_build', 'contracts', 'types', 'examples', 'grammar', 'ir']
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.